+++ templates/invoices/new_invoice.liquid
@@ -165,7 +165,6 @@
             row.insertCell().innerHTML = `<input type="text" id="price${rowID}" onChange="updateTotal()">`
             row.insertCell().innerHTML = `<input type="text" id="qty${rowID}" onChange="updateTotal()">`
             row.insertCell().innerHTML = `<button onClick="removeRow(${rowID})">x</button>`
-
         }
 
         function preview(){
@@ -366,11 +365,9 @@
         <label for="ID">Record ID</label>
         <input type="text" name="ID" id="ID" disabled> <br>
         <label for="creator">Created By:</label>
-        <input type="text" id="creator" disabled><br>
+        <input type="text" id="creator" disabled> <button onclick="getCreatorUser()">Get User</button><br>
         <label for="approver">Approved By:</label>
-        <input type="text" id="approver" {% if status!="pending" or type == "sga_budget"%}
-            disabled
-        {% endif %}><br>
+        <input type="text" id="approver" disabled><button onclick="getApproverUser()">Get User</button><br>
         <label for="createDate">Creation Date</label>
         <input type="date" name="createDate" id="createDate">
     </fieldset>
