--- texput.log
@@ -1,21 +1,0 @@
-This is pdfTeX, Version 3.141592653-2.6-1.40.24 (TeX Live 2022/Debian) (preloaded format=pdflatex 2024.1.9)  9 JAN 2024 22:46
-entering extended mode
- restricted \write18 enabled.
- %&-line parsing enabled.
-**
-
-! Emergency stop.
-<*> 
-    
-End of file on the terminal!
-
- 
-Here is how much of TeX's memory you used:
- 7 strings out of 476091
- 113 string characters out of 5794082
- 1849330 words of memory out of 5000000
- 20488 multiletter control sequences out of 15000+600000
- 512287 words of font info for 32 fonts, out of 8000000 for 9000
- 1141 hyphenation exceptions out of 8191
- 0i,0n,0p,9b,6s stack positions out of 10000i,1000n,20000p,200000b,200000s
-!  ==> Fatal error occurred, no output PDF file produced!
